COMPASS-697: Backport COMPASS-666: Keyboard Enter on Create Collection Modal sometimes cause page refresh to 1.5-releases (#758)
* COMPASS-697: Backport COMPASS-666: Keyboard Enter on forms cause page refresh (#753)

NOTE: Due to merge conflicts in our rapidly evolving functional test suite since creating the 1.5.x branch, ‘data-test-id’ values and the actual code fix has been backported and manually tested, but the functional test changes have not been back ported.

* COMPASS 443: Standardise mongodb-runner CI (#633)

* Standardise mongodb-runner CI

https://github.com/mongodb-js/connection-model/pull/138

NOTE: `npm run check && npm run test-unit` passes for me locally on macOS, but the functional tests `npm test ` still error with results:

  177 passing (50s)
  2 pending
  10 failing

  1) Compass #spectron when working with the application when working with collections when selecting a collection renders the sample collection in the title:
     Error: element (button[data-hook=start-button]) still not visible after 5000ms
      at elements("button[data-hook=start-button]") - isVisible.js:54:17
      at isVisible("button[data-hook=start-button]") - waitForVisible.js:37:22

  2) Compass #spectron when working with the application when working with collections when selecting a collection renders the schema tab:
     Error: An element could not be located on the page using the given search parameters.
      at element("li#SCHEMA") - click.js:12:17
      at elementIdDisplayed("0.9192226665154364-7") - isVisible.js:71:55

  3) Compass #spectron when working with the application when working with collections when selecting a collection when applying a filter the text in refine bar matches query:
     Error: Promise was rejected with the following reason: Error: An element could not be located on the page using the given search parameters ("input#refine_input").
      at elements("input#refine_input") - getValue.js:18:17
      at getValue("input#refine_input") - waitForValue.js:37:22

  4) Compass #spectron when working with the application when working with collections when selecting a collection when applying a filter samples the matching documents:
     Error: element (div.sampling-message) still not visible after 15000ms
      at elements("div.sampling-message") - isVisible.js:54:17
      at isVisible("div.sampling-message") - waitForVisible.js:37:22

  5) Compass #spectron when working with the application when working with collections when selecting a collection when applying a filter updates the schema view:
     Error: An element could not be located on the page using the given search parameters ("li.bubble code.selectable").
      at elements("li.bubble code.selectable") - getText.js:18:17
      at getText("li.bubble code.selectable") - compass-functional.test.js:120:16

  6) Compass #spectron when working with the application when working with collections when selecting a collection when applying a filter filters out non-matches from the document list:
     Error: An element could not be located on the page using the given search parameters.
      at element("li#DOCUMENTS") - click.js:12:17
      at elementIdDisplayed("0.9192226665154364-7") - isVisible.js:71:55

  7) Compass #spectron when working with the application when working with collections when selecting a collection when applying a filter includes documents that match the filter:
     Error: An element could not be located on the page using the given search parameters ("div.element-value-is-string").
      at elements("div.element-value-is-string") - getText.js:18:17
      at getText("div.element-value-is-string") - compass-functional.test.js:138:16

  8) Compass #spectron when working with the application when working with collections when working in the documents tab when viewing documents renders the documents in the list:
     Error: An element could not be located on the page using the given search parameters.
      at element("li#DOCUMENTS") - click.js:12:17
      at elementIdDisplayed("0.9192226665154364-7") - isVisible.js:71:55

  9) Compass #spectron when working with the application when working with collections when working in the explain tab when viewing the explain tab renders the stages:
     Error: An element could not be located on the page using the given search parameters.
      at element("li#EXPLAIN_PLAN") - click.js:12:17
      at elementIdDisplayed("0.9192226665154364-7") - isVisible.js:71:55

  10) Compass #spectron when working with the application when working with collections when working in the indexes tab when viewing the indexes tab renders the indexes:
     Error: An element could not be located on the page using the given search parameters.
      at element("li#INDEXES") - click.js:12:17
      at elementIdDisplayed("0.9192226665154364-7") - isVisible.js:71:55

* Go back to the full pretest and post test hooks

Annoying as we can’t test multiple topologies, but probably a decent idea for Compass itself as we should delegate as much as possible to each public mongodb-js dependency.

Note: Needed as otherwise TravisCI hangs for 10 minutes then kills the job even though it completes with exit code 0, i.e. successfully.
e.g. https://travis-ci.com/10gen/compass/jobs/57691028

* Drop lts/boron

We can add it in when we decide we actually want to switch, as our private Travis seems to be a little under-powered at this time (build queues backing up).

* Standardise on no space

It’s that way on master, just too lazy to mine git log to cherry-pick the change that did it somewhere between 1.5 and 1.7, to pass code review comments.
https://github.com/10gen/compass/pull/758

* Drop data-test-ids we do not need on 1.5-releases

+++ src/internal-packages/database/lib/components/drop-collection-dialog.jsx
@@ -62,8 +62,17 @@ class DropCollectionDialog extends React.Component {
 
   /**
    * Initiate the attempt to drop a database.
+   * @param {Object} evt - The event object
    */
-  onDropCollectionButtonClicked() {
+  onDropCollectionButtonClicked(evt) {
+    evt.preventDefault();
+    evt.stopPropagation();
+
+    // prevent drop of collection if names don't match
+    if (this.state.confirmName !== this.state.name) {
+      return;
+    }
+
     this.setState({ inProgress: true, error: false, errorMessage: '' });
     Actions.dropCollection(this.state.databaseName, this.state.name);
   }
@@ -111,7 +120,9 @@ class DropCollectionDialog extends React.Component {
               to drop
             </p>
           </div>
-          <form>
+          <form
+            onSubmit={this.onDropCollectionButtonClicked.bind(this)}
+          >
             <div className="form-group">
               <input
                 type="text"
